+++ example/lib/main.dart
@@ -13,6 +13,7 @@ class MyApp extends StatelessWidget {
 
   // This widget is the root of your application.
   final formKey = GlobalKey<FormState>();
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -69,10 +70,8 @@ class MyApp extends StatelessWidget {
                         text: "New York City",
                       ),
                       initData: [
-                        DropDownModel<String>(
-                            key: "1", value: "1", text: "New York City"),
-                        DropDownModel<String>(
-                            key: "2", value: "2", text: "Los Angeles"),
+                        DropDownModel<String>(key: "1", value: "1", text: "New York City"),
+                        DropDownModel<String>(key: "2", value: "2", text: "Los Angeles"),
                       ],
                     ),
                     const SizedBox(
@@ -98,10 +97,8 @@ class MyApp extends StatelessWidget {
                         log("SELECTED VALUE ${value?.value}");
                       },
                       initData: [
-                        DropDownModel<String>(
-                            key: "1", value: "1", text: "New York City"),
-                        DropDownModel<String>(
-                            key: "2", value: "2", text: "Los Angeles"),
+                        DropDownModel<String>(key: "1", value: "1", text: "New York City"),
+                        DropDownModel<String>(key: "2", value: "2", text: "Los Angeles"),
                       ],
                     ),
                     const SizedBox(
@@ -149,31 +146,13 @@ class MyApp extends StatelessWidget {
                             disabled: true,
                             disabledText: "Sold out",
                           ),
-                          // DropDownModel<String>(
-                          //     key: "3", value: "3", text: "Chicago"),
-                          // DropDownModel<String>(
-                          //     key: "4", value: "4", text: "Houston"),
-                          // DropDownModel<String>(
-                          //     key: "5", value: "5", text: "Phoenix"),
-                          // DropDownModel<String>(
-                          //     key: "3", value: "3", text: "Philadelphia"),
-                          // DropDownModel<String>(
-                          //     key: "4", value: "4", text: "San Antonio"),
-                          // DropDownModel<String>(
-                          //     key: "5", value: "5", text: "San Diego"),
-                          // DropDownModel<String>(
-                          //     key: "3", value: "3", text: "Dallas"),
-                          // DropDownModel<String>(
-                          //     key: "4", value: "4", text: "Austin"),
-                          // DropDownModel<String>(
-                          //     key: "5", value: "5", text: "Texas"),
                         ];
                       }(),
                     ),
                     const SizedBox(
                       height: 8,
                     ),
-                    DfSearchableSingleSelectDropdown<String>(
+                    DfSearchableSingleSelectDropdown<int>(
                       validator: (obj) {
                         return "This is a really long validation error, an this is how it will be displayed...";
                       },
@@ -182,24 +161,137 @@ class MyApp extends StatelessWidget {
                       labelText: "Single Select",
                       selectorDecoration: const SingleSelectorDecoration(
                         selectedItemIcon: Icon(Icons.circle),
+                        searchTextStyle: TextStyle(fontSize: 14)
                       ),
                       onOptionSelected: (value) {
                         log("SELECTED VALUE ${value?.value}");
                       },
-                      selectedValue: DropDownModel<String>(
-                          key: "3", value: "3", text: "Chicago"),
-                      initData: [
-                        DropDownModel<String>(
-                            key: "1", value: "1", text: "New York City"),
-                        DropDownModel<String>(
-                            key: "2", value: "2", text: "Los Angeles"),
-                        DropDownModel<String>(
-                            key: "3", value: "3", text: "Chicago"),
-                        DropDownModel<String>(
-                            key: "4", value: "4", text: "Houston"),
-                        DropDownModel<String>(
-                            key: "5", value: "5", text: "Phoenix"),
+                      nestedInitData: [
+                        DropDownNestedModel(
+                          title: "Bosna i Hercegovina",
+                          children: [
+                            DropDownNestedModel(
+                              title: "Republika Srpska",
+                              children: [
+                                DropDownNestedModel(
+                                  title: "Sarajevsko-Zenička regija",
+                                  values: [
+                                    DropDownModel(key: "sarajevo", text: "Sarajevo", value: 1),
+                                    DropDownModel(key: "zenica", text: "Zenica", value: 2),
+                                  ],
+                                ),
+                                DropDownNestedModel(
+                                  title: "Banjalučka regija",
+                                  values: [
+                                    DropDownModel(key: "banjaluka", text: "Banja Luka", value: 3),
+                                    DropDownModel(key: "prijedor", text: "Prijedor", value: 4),
+                                  ],
+                                ),
+                              ],
+                            ),
+                            DropDownNestedModel(
+                              title: "Federacija BiH",
+                              children: [
+                                DropDownNestedModel(
+                                  title: "Hercegovačka regija",
+                                  values: [
+                                    DropDownModel(key: "mostar", text: "Mostar", value: 5),
+                                    DropDownModel(key: "trebinje", text: "Trebinje", value: 6),
+                                  ],
+                                ),
+                                DropDownNestedModel(
+                                  title: "Tuzlanska regija",
+                                  values: [
+                                    DropDownModel(key: "tuzla", text: "Tuzla", value: 7),
+                                    DropDownModel(key: "brcko", text: "Brčko", value: 8),
+                                  ],
+                                ),
+                                DropDownNestedModel(
+                                  title: "Unsko-Sanska regija",
+                                  values: [
+                                    DropDownModel(key: "bihac", text: "Bihać", value: 9),
+                                    DropDownModel(key: "cazin", text: "Cazin", value: 10),
+                                  ],
+                                ),
+                              ],
+                            ),
+                          ],
+                        ),
+                        DropDownNestedModel(
+                          title: "Srbija",
+                          children: [
+                            DropDownNestedModel(
+                              title: "Beogradski region",
+                              values: [
+                                DropDownModel(key: "beograd", text: "Beograd", value: 1),
+                              ],
+                            ),
+                            DropDownNestedModel(
+                              title: "Vojvodina",
+                              children: [
+                                DropDownNestedModel(
+                                  title: "Severnobački okrug",
+                                  values: [
+                                    DropDownModel(key: "subotica", text: "Subotica", value: 2),
+                                  ],
+                                ),
+                                DropDownNestedModel(
+                                  title: "Južnobački okrug",
+                                  values: [
+                                    DropDownModel(key: "novisad", text: "Novi Sad", value: 3),
+                                    DropDownModel(key: "sombor", text: "Sombor", value: 4),
+                                  ],
+                                ),
+                              ],
+                            ),
+                            DropDownNestedModel(
+                              title: "Centralna Srbija",
+                              children: [
+                                DropDownNestedModel(
+                                  title: "Šumadijski okrug",
+                                  values: [
+                                    DropDownModel(key: "kragujevac", text: "Kragujevac", value: 5),
+                                  ],
+                                ),
+                                DropDownNestedModel(
+                                  title: "Nišavski okrug",
+                                  values: [
+                                    DropDownModel(key: "nis", text: "Niš", value: 6),
+                                    DropDownModel(key: "pirot", text: "Pirot", value: 7),
+                                  ],
+                                ),
+                                DropDownNestedModel(
+                                  title: "Pomoravski okrug",
+                                  values: [
+                                    DropDownModel(key: "jagodina", text: "Jagodina", value: 8),
+                                    DropDownModel(key: "cuprija", text: "Ćuprija", value: 9),
+                                  ],
+                                ),
+                              ],
+                            ),
+                            DropDownNestedModel(
+                              title: "Zapadna Srbija",
+                              values: [
+                                DropDownModel(key: "uzice", text: "Užice", value: 10),
+                              ],
+                            ),
+                          ],
+                        ),
                       ],
+                      // selectedValue: DropDownModel<String>(
+                      //     key: "3", value: "3", text: "Chicago"),
+                      // initData: [
+                      //   DropDownModel<String>(
+                      //       key: "1", value: "1", text: "New York City"),
+                      //   DropDownModel<String>(
+                      //       key: "2", value: "2", text: "Los Angeles"),
+                      //   DropDownModel<String>(
+                      //       key: "3", value: "3", text: "Chicago"),
+                      //   DropDownModel<String>(
+                      //       key: "4", value: "4", text: "Houston"),
+                      //   DropDownModel<String>(
+                      //       key: "5", value: "5", text: "Phoenix"),
+                      // ],
                     ),
                     const SizedBox(
                       height: 8,
@@ -216,40 +308,26 @@ class MyApp extends StatelessWidget {
                         borderColor: Colors.amber,
                         outlineBorderColor: Colors.blue,
                       ),
-                      selectorDecoration: const MultiSelectorDecoration(
-                          selectedItemColor: Colors.orange),
+                      selectorDecoration: const MultiSelectorDecoration(selectedItemColor: Colors.orange),
                       onOptionSelected: (value) {
                         log("SELECTED VALUE ${value.length}");
                       },
                       selectedValues: [
-                        DropDownModel<String>(
-                            key: "1", value: "1", text: "New York City"),
-                        DropDownModel<String>(
-                            key: "2", value: "2", text: "Los Angeles"),
+                        DropDownModel<String>(key: "1", value: "1", text: "New York City"),
+                        DropDownModel<String>(key: "2", value: "2", text: "Los Angeles"),
                       ],
                       initData: [
-                        DropDownModel<String>(
-                            key: "1", value: "1", text: "New York City"),
-                        DropDownModel<String>(
-                            key: "2", value: "2", text: "Los Angeles"),
-                        DropDownModel<String>(
-                            key: "3", value: "3", text: "Chicago"),
-                        DropDownModel<String>(
-                            key: "4", value: "4", text: "Houston"),
-                        DropDownModel<String>(
-                            key: "5", value: "5", text: "Phoenix"),
-                        DropDownModel<String>(
-                            key: "6", value: "3", text: "Philadelphia"),
-                        DropDownModel<String>(
-                            key: "7", value: "4", text: "San Antonio"),
-                        DropDownModel<String>(
-                            key: "8", value: "5", text: "San Diego"),
-                        DropDownModel<String>(
-                            key: "9", value: "3", text: "Dallas"),
-                        DropDownModel<String>(
-                            key: "10", value: "4", text: "Austin"),
-                        DropDownModel<String>(
-                            key: "11", value: "5", text: "Texas"),
+                        DropDownModel<String>(key: "1", value: "1", text: "New York City"),
+                        DropDownModel<String>(key: "2", value: "2", text: "Los Angeles"),
+                        DropDownModel<String>(key: "3", value: "3", text: "Chicago"),
+                        DropDownModel<String>(key: "4", value: "4", text: "Houston"),
+                        DropDownModel<String>(key: "5", value: "5", text: "Phoenix"),
+                        DropDownModel<String>(key: "6", value: "3", text: "Philadelphia"),
+                        DropDownModel<String>(key: "7", value: "4", text: "San Antonio"),
+                        DropDownModel<String>(key: "8", value: "5", text: "San Diego"),
+                        DropDownModel<String>(key: "9", value: "3", text: "Dallas"),
+                        DropDownModel<String>(key: "10", value: "4", text: "Austin"),
+                        DropDownModel<String>(key: "11", value: "5", text: "Texas"),
                       ],
                     ),
                     ElevatedButton(