+++ lib/df_searchable_single_select_dropdown.dart
@@ -27,6 +27,8 @@ class DfSearchableSingleSelectDropdown<T> extends BaseDropdown<T> {
     super.dropdownType = DropdownType.expandable,
     super.disabled = false,
     this.closeDropdownOnSelection = true,
+    super.nestedInitData,
+    super.asyncNestedInitData,
     super.asyncInitData,
   }) : assert(initData.length == 0 || asyncInitData == null,
             "initData and asyncInitData cannot be provided at the same time");
@@ -50,8 +52,10 @@ class DfSearchableSingleSelectDropdown<T> extends BaseDropdown<T> {
         onOptionSelected: onOptionSelected,
         validator: validator,
         onSearch: onSearch,
+        asyncNestedInitData: asyncNestedInitData,
         selectorMaxHeight: selectorDecoration?.maxHeight,
         closeDropdownOnSelection: closeDropdownOnSelection,
+        nestedInitData: nestedInitData??[],
         context: context,
       ),
       child: _Dropdown<T>(