+++ lib/models/models.dart
@@ -4,3 +4,4 @@ export './multi_selector_decoration.dart';
 export './simple_selector_decoration.dart';
 export './single_selector_decoration.dart';
 export './searchable_selector_base_decoration.dart';
+export './drop_down_nested_model.dart';