+++ lib/providers/searchable_multi_select_dropdown_provider.dart
@@ -9,6 +9,8 @@ class SearchableMultiSelectDropdownProvider<T> extends BaseDropdownProvider<T> {
     this.onOptionSelected,
     this.onSearch,
     super.initData,
+    super.nestedInitData,
+    super.asyncNestedInitData,
     this.multiSelectValidator,
     this.selectorMaxHeight,
     this.displayResultsCount,