+++ lib/providers/simple_dropdown_provider.dart
@@ -7,6 +7,8 @@ class SimpleDropdownProvider<T> extends BaseDropdownProvider<T> {
     this.onOptionSelected,
     this.maxHeight,
     super.initData,
+    super.asyncNestedInitData,
+    super.nestedInitData,
     super.validator,
     required super.asyncInitData,
     required super.context,